refactor(studio): share chain routing policy
