+++ studio/index.ts
@@ -4,5 +4,6 @@ export * from "./platform";
 export * from "./stores/runtime";
 export * from "./lib/promptCycler";
 export * from "./lib/sourceFit";
+export * from "./lib/chainRouting";
 export * from "./lib/modelDisplay";
 export * from "./lib/sequence";
